requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-countries==7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
names==0.3.0
packaging==20.8
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
rope==0.18.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
